Fix regression in Sending JSON reward
I broke the JSON reward path in the UTF fix due to a copy-and-paste error, and on top of that the test that should have caught it did not actually have any call into the test code, it just set up the test with running it.

On top of that, test.cmd was never updated to run the .NET unit tests.

This is all fixed by this change.